Separate teensy4-bsp pins API into new crate
