+++ Cargo.toml
@@ -21,6 +21,10 @@ teensy4-fcb = { path = "teensy4-fcb" }
 version = "0.4.0"
 features = ["imxrt1062", "rt"]
 
+[dependencies.teensy4-pins]
+version = "0.1"
+path = "teensy4-pins"
+
 # Tied to "systick" feature, since
 # SysTick implements a blocking delay trait
 [dependencies.embedded-hal]
@@ -52,6 +56,7 @@ panic-halt = "0.2.0"
 members = [
     "cortex-m-rt-patch",
     "teensy4-fcb",
+    "teensy4-pins",
     "teensy4-rt",
     "teensy4-usb-sys",
 ]
